+++ src/hd98.c
@@ -165,7 +165,7 @@ HD98_Material *hd98_halm_dragon_1998_new_default() {
 }
 
 void hd98_global_update(size_t n, double const *delta_eps, double const *eps1,
-                        double const *omega1, uint8_t const *phase,
+                        double const *omega1, size_t const *phase,
                         HD98_Material **mat, double *sig2, double *omega2,
                         double *C2) {
   double const *delta_eps_i = delta_eps;
